Issue #2087117 by jyee: Added hook_block_configure in block configure form.
- fgm: split old-style test to PSR4.
